+++ packages/interactions/__tests__/lex-v2/apis/send.test.ts
@@ -2,23 +2,23 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { v4 as uuid } from 'uuid';
-import { lexProvider } from '../../../src/lex-v1/AWSLexProvider';
-import { sendMessage } from '../../../src/lex-v1/apis';
-import { generateRandomLexV1Config } from '../../testUtils/randomConfigGeneration.test';
-import { resolveBotConfig } from '../../../src/lex-v1/utils';
+import { lexProvider } from '../../../src/lex-v2/AWSLexV2Provider';
+import { send } from '../../../src/lex-v2/apis';
+import { generateRandomLexV2Config } from '../../testUtils/randomConfigGeneration.test';
+import { resolveBotConfig } from '../../../src/lex-v2/utils';
 import { InteractionsError } from '../../../src/errors/InteractionsError';
 
-jest.mock('../../../src/lex-v1/AWSLexProvider');
-jest.mock('../../../src/lex-v1/utils');
+jest.mock('../../../src/lex-v2/AWSLexV2Provider');
+jest.mock('../../../src/lex-v2/utils');
 
-describe('Interactions LexV1 API: sendMessage', () => {
-	const v1BotConfig = generateRandomLexV1Config();
+describe('Interactions LexV2 API: send', () => {
+	const v2BotConfig = generateRandomLexV2Config();
 
 	const mockLexProvider = lexProvider.sendMessage as jest.Mock;
 	const mockResolveBotConfig = resolveBotConfig as jest.Mock;
 
 	beforeEach(() => {
-		mockResolveBotConfig.mockReturnValue(v1BotConfig);
+		mockResolveBotConfig.mockReturnValue(v2BotConfig);
 	});
 
 	afterEach(() => {
@@ -28,14 +28,14 @@ describe('Interactions LexV1 API: sendMessage', () => {
 
 	it('invokes provider sendMessage API', async () => {
 		const message = uuid();
-		await sendMessage(v1BotConfig.name, message);
+		await send(v2BotConfig.name, message);
 		expect(mockLexProvider).toBeCalledTimes(1);
-		expect(mockLexProvider).toBeCalledWith(v1BotConfig, message);
+		expect(mockLexProvider).toBeCalledWith(v2BotConfig, message);
 	});
 
 	it('rejects when bot config does not exist', async () => {
 		mockResolveBotConfig.mockReturnValue(undefined);
-		await expect(sendMessage(v1BotConfig.name, uuid())).rejects.toBeInstanceOf(
+		await expect(send(v2BotConfig.name, uuid())).rejects.toBeInstanceOf(
 			InteractionsError
 		);
 	});
