+++ packages/interactions/src/index.ts
@@ -1,5 +1,11 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './types';
+export {
+	InteractionsMessage,
+	InteractionsResponse,
+	InteractionsVoiceMessage,
+	InteractionsTextMessage,
+	CompletionCallback,
+} from './types';
 export { Interactions } from './lex-v2';
