+++ packages/interactions/src/lex-v1/apis/index.ts
@@ -1,5 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export { sendMessage } from './sendMessage';
+export { send } from './send';
 export { onComplete } from './onComplete';
