+++ packages/interactions/src/lex-v1/index.ts
@@ -1,10 +1,10 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { sendMessage, onComplete } from './apis';
+import { send, onComplete } from './apis';
 import { IInteractions } from '../types/Interactions';
 
 export const Interactions: IInteractions = {
-	send: sendMessage,
+	send,
 	onComplete,
 };
