+++ packages/interactions/src/lex-v2/apis/send.ts
@@ -2,14 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { InteractionsMessage, InteractionsResponse } from '../../types';
+import { lexProvider } from '../AWSLexV2Provider';
 import { resolveBotConfig } from '../utils';
-import { lexProvider } from '../AWSLexProvider';
 import {
 	assertValidationError,
 	InteractionsValidationErrorCode,
 } from '../../errors';
 
-export const sendMessage = async (
+export const send = async (
 	botName: string,
 	message: string | InteractionsMessage
 ): Promise<InteractionsResponse> => {
