+++ packages/interactions/src/utils/index.ts
@@ -1,5 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './utils';
-export * from './commonUtils';
+export { convert, base64ToArrayBuffer, gzipDecompressToString } from './utils';
+export { unGzipBase64AsJson } from './commonUtils';
